requirements.txt
openai==1.3.0
transformers==4.35.2
torch==2.1.1
numpy==1.26.2
pandas==2.1.3
tqdm==4.66.1
requests==2.31.0
python-dotenv==1.0.0
sqlparse==0.4.4
psycopg2-binary==2.9.9
typing-extensions==4.8.0
pydantic==2.5.2